[Fix] 배포 전 오류 수정
