+++ src/components/mypage/order/OrderPaymentDetails.tsx
@@ -18,7 +18,9 @@ const OrderPaymentDetails = ({ item }: PaymentDetailsProps) => {
       </li>
       <li>
         <span className="text-gray-500 mr-1">결제 일시 :</span>
-        <span>{formatPaymentDate(item?.created_at!)}</span>
+        <span>
+          {item?.created_at ? formatPaymentDate(item.created_at) : "-"}
+        </span>
       </li>
       <li>
         <span className="text-gray-500 mr-1">주문 상태 :</span>
